feat: change `block_height` to u32

+++ crates/rust-client/src/notes/import.rs
@@ -100,7 +100,7 @@ impl<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator> Client
                 NoteStatus::Consumed { consumer_account_id: None, block_height }
             } else {
                 NoteStatus::Committed {
-                    block_height: inclusion_proof.location().block_num() as u64,
+                    block_height: inclusion_proof.location().block_num(),
                 }
             };
 
@@ -143,7 +143,7 @@ impl<N: NodeRpcClient, R: FeltRng, S: Store, A: TransactionAuthenticator> Client
             NoteStatus::Consumed { consumer_account_id: None, block_height }
         } else {
             NoteStatus::Committed {
-                block_height: inclusion_proof.location().block_num() as u64,
+                block_height: inclusion_proof.location().block_num(),
             }
         };
